fix: align user.id PK type and org_id FK type with entity declarations
- Change user.id from BIGSERIAL to SERIAL in baseline migration to match User entity @PrimaryGeneratedColumn() (integer/SERIAL)
- Change OrgInventoryItem.orgId column type from bigint to integer to match Organization.id (SERIAL = INTEGER) and the migration DDL

+++ backend/src/modules/org-inventory/entities/org-inventory-item.entity.ts
@@ -30,7 +30,7 @@ export class OrgInventoryItem {
   @PrimaryGeneratedColumn('uuid')
   id!: string;
 
-  @Column({ name: 'org_id', type: 'bigint' })
+  @Column({ name: 'org_id', type: 'integer' })
   @Index()
   orgId!: number;
